Create modify_number function
This function handles the logic of the caesar cipher; it receives two
parameters: a number that represents the character and another one
that is the shift factor.

When the number is not a letter, it just returns the character; in case
its a letter and it passes the last letter with the shift factor, it
calculate which letter should be and finally when the num does not
pass the last letter, it just adds the shift factor. At the end it
returns the character.

In the caesar cipher with the each method it applies this function and
appends the results in the modified_string variable and when its done
it returns the result.